Integrated ACRA

+++ WAKiMail/src/net/rdrei/android/wakimail/task/MailSyncTask.java
@@ -1,7 +1,5 @@
 package net.rdrei.android.wakimail.task;
 
-import java.util.Iterator;
-
 import net.rdrei.android.wakimail.R;
 import net.rdrei.android.wakimail.data.MailTable;
 import net.rdrei.android.wakimail.wak.Mail;
@@ -60,7 +58,7 @@ public class MailSyncTask extends RdreiAsyncTask<Integer> {
 	protected void onSuccess(Integer result) throws Exception {
 		super.onSuccess(result);
 		
-		if (result > 0) {
+		if (result.intValue() > 0) {
 			showResultToast(result);
 		}
 	}
